services-python/vision-server/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
flask==3.0.0
flask-cors==4.0.0
anthropic>=0.18.0
# LM Studio API (OpenAI-compatible)
openai>=1.0.0

# Validation and Review Tools
matplotlib>=3.8.0
pillow>=10.0.0
requests>=2.31.0  # For API testing and validation
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
onnx>=1.15.0
onnxruntime>=1.17.0
